refactor: extract pipeline orchestration from cmd/gotest into internal/gotestrunner
