+++ internal/gotestrunner/export_test.go
@@ -1,10 +1,22 @@
 package gotestrunner
 
+import "encoding/json"
+
 type ExportOverlayJSON = overlayJSON
+type ExportFixtureStateEntry = fixtureStateEntry
 
 var ExportBuildSuiteCmd = buildSuiteCmd
 var ExportReadTeardownBudget = readTeardownBudget
 var ExportSplitTopLevelOr = splitTopLevelOr
 var ExportSuiteRunFilter = suiteRunFilter
+var ExportAssignCoverProfiles = assignCoverProfiles
+var ExportBuildExtraEnv = buildExtraEnv
 
 var SetBuildProcessGroup = setBuildProcessGroup
+
+func ExportNewSharedFixtureProcess(sharedDir string, state map[string]json.RawMessage) *SharedFixtureProcess {
+	return &SharedFixtureProcess{
+		sharedDir: sharedDir,
+		state:     state,
+	}
+}
